Revert "Revert "aa""
This reverts commit 811ba7a98e14076ae14ce2de21705d830bb56f06.

+++ src/main/java/com/newer/sc/common/entity/Teacher.java
@@ -12,7 +12,7 @@ import org.springframework.stereotype.Component;
 @Component
 public class Teacher {
 	
-	int tid;
+	Long tid;
 	
 	String tname;
 	
@@ -28,11 +28,11 @@ public class Teacher {
 	
 	College college;
 
-	public int getTid() {
+	public Long getTid() {
 		return tid;
 	}
 
-	public void setTid(int tid) {
+	public void setTid(Long tid) {
 		this.tid = tid;
 	}
